requirements.txt
rembg>=2.0.50
Pillow>=10.0.0
onnxruntime>=1.15.0

# Pour le développement et les tests
pytest>=7.0.0
